added tests for CreditCard class
